Create AuthDomainMultiGroupSpec
fix
