+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/security/AuthDomainMultiGroupsSpec.scala
@@ -2,28 +2,18 @@ package org.broadinstitute.dsde.firecloud.test.security
 
 import org.broadinstitute.dsde.firecloud.fixture.UserFixtures
 import org.broadinstitute.dsde.firecloud.page.workspaces.summary.WorkspaceSummaryPage
-import org.broadinstitute.dsde.firecloud.test.Tags
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.config.{Credentials, UserPool}
 import org.broadinstitute.dsde.workbench.fixture.{BillingFixtures, GroupFixtures, TestReporterFixture, WorkspaceFixtures}
-import org.broadinstitute.dsde.workbench.service.{AclEntry, Orchestration, WorkspaceAccessLevel}
 import org.broadinstitute.dsde.workbench.service.Orchestration.billing.BillingProjectRole
 import org.broadinstitute.dsde.workbench.service.test.{CleanUp, WebBrowserSpec}
+import org.broadinstitute.dsde.workbench.service.{AclEntry, Orchestration, WorkspaceAccessLevel}
 import org.scalatest._
 import org.scalatest.concurrent.Eventually
 import org.scalatest.time.{Millis, Seconds, Span}
 
-/*
- * This test SHOULD be able to run with ParallelTestExecution. However, Rawls
- * currently has database deadlock problems when creating and deleting managed
- * groups because they involve 3 operations across 2 tables. My initial
- * attempt to retry the deadlock failed because Rawls also creates the groups
- * in Google inside the transaction; attempts to retry the transaction result
- * in 409 Conflict errors from Google.
- *
- * TODO: Fix Rawls group creation/deletion and run these tests in parallel
- */
-class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ with Matchers with CleanUp
+
+class AuthDomainMultiGroupsSpec extends FreeSpec with ParallelTestExecution with Matchers with CleanUp
   with WebBrowserSpec with WorkspaceFixtures with Eventually with BillingFixtures with GroupFixtures with UserFixtures
   with TestReporterFixture {
 
@@ -44,190 +34,11 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
     error should include("does not exist")
   }
 
+  // MULTI-GROUP AUTH DOMAIN TESTS
+
   "A workspace with an authorization domain" - {
-    "with one group inside of it" - {
-      "can be created" in {
-        val user = UserPool.chooseAuthDomainUser
-        implicit val authToken: AuthToken = authTokenDefault
-        withGroup("AuthDomain", List(user.email)) { authDomainName =>
-          withCleanUp {
-            withCleanBillingProject(user) { projectName =>
-              withWebDriver { implicit driver =>
-                withSignIn(user) { listPage =>
-                  val workspaceName = "AuthDomainSpec_create_" + randomUuid
-                  register cleanUp api.workspaces.delete(projectName, workspaceName)(user.makeAuthToken())
-                  val workspaceSummaryPage = listPage.createWorkspace(projectName, workspaceName, Set(authDomainName))
-
-                  eventually {
-                    workspaceSummaryPage.readAuthDomainGroups should include(authDomainName)
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-      "can be cloned and retain the auth domain" taggedAs Tags.SmokeTest in {
-        val user = UserPool.chooseAuthDomainUser
-        implicit val authToken: AuthToken = authTokenDefault
-        withGroup("AuthDomain", List(user.email)) { authDomainName =>
-          withCleanBillingProject(defaultUser) { projectName =>
-            Orchestration.billing.addUserToBillingProject(projectName, user.email, BillingProjectRole.User)
-            withWorkspace(projectName, "AuthDomainSpec_share", Set(authDomainName), List(AclEntry(user.email, WorkspaceAccessLevel.Reader))) { workspaceName =>
-              withCleanUp {
-                withWebDriver { implicit driver =>
-                  withSignIn(user) { listPage =>
-                    val summaryPage = listPage.enterWorkspace(projectName, workspaceName)
-
-                    val cloneWorkspaceName = workspaceName + "_clone"
-                    val cloneModal = summaryPage.clickCloneButton()
-                    eventually {
-                      cloneModal.readLockedAuthDomainGroups() should contain(authDomainName)
-                    }
-
-                    register cleanUp api.workspaces.delete(projectName, cloneWorkspaceName)(authTokenDefault)
-
-                    val cloneSummaryPage = cloneModal.cloneWorkspace(projectName, cloneWorkspaceName)
-                    eventually {
-                      cloneSummaryPage.validateWorkspace shouldEqual true
-                    }
-                    eventually {
-                      cloneSummaryPage.readAuthDomainGroups should include(authDomainName)
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-      "when the user is not inside of the group" - {
-        "when the workspace is shared with them" - {
-          "can be seen but is not accessible" in {
-            val user = UserPool.chooseStudent
-            implicit val authToken: AuthToken = authTokenDefault
-            withGroup("AuthDomain") { authDomainName =>
-              withCleanBillingProject(defaultUser) { projectName =>
-                withWorkspace(projectName, "AuthDomainSpec_reject", Set(authDomainName), List(AclEntry(user.email, WorkspaceAccessLevel.Reader))) { workspaceName =>
-                  withWebDriver { implicit driver =>
-                    withSignIn(user) { workspaceListPage =>
-                      workspaceListPage.clickWorkspaceLink(projectName, workspaceName)
-                      eventually { workspaceListPage.showsRequestAccessModal shouldEqual true }
-                      workspaceListPage.validateLocation()
-                      // TODO: add assertions for the new "request access" modal
-                      // TODO: end test somewhere we can access the sign out button
-
-                      // close "request access" Modal
-                      workspaceListPage.closeModal()
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        "when the workspace is not shared with them" - {
-          "cannot be seen and is not accessible" in {
-            val user = UserPool.chooseStudent
-            implicit val authToken: AuthToken = authTokenDefault
-            withGroup("AuthDomain") { authDomainName =>
-              withCleanBillingProject(defaultUser) { projectName =>
-                withWorkspace(projectName, "AuthDomainSpec", Set(authDomainName)) { workspaceName =>
-                  withWebDriver { implicit driver =>
-                    withSignIn(user) { workspaceListPage =>
-                      eventually {
-                        workspaceListPage.hasWorkspace(projectName, workspaceName) shouldEqual false
-                      }
-
-                      val workspaceSummaryPage = new WorkspaceSummaryPage(projectName, workspaceName).open
-                      checkWorkspaceFailure(workspaceSummaryPage, projectName, workspaceName)
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-      "when the user is inside of the group" - {
-        "when the workspace is shared with them" - {
-          "can be seen and is accessible" in {
-
-            val user = UserPool.chooseAuthDomainUser
-            implicit val authToken: AuthToken = authTokenDefault
-            withGroup("AuthDomain", List(user.email)) { authDomainName =>
-              withCleanBillingProject(defaultUser) { projectName =>
-                withWorkspace(projectName, "AuthDomainSpec_share", Set(authDomainName), List(AclEntry(user.email, WorkspaceAccessLevel.Reader))) { workspaceName =>
-                  withWebDriver { implicit driver =>
-                    withSignIn(user) { listPage =>
-                      val summaryPage = listPage.enterWorkspace(projectName, workspaceName)
-                      eventually {
-                        summaryPage.readAuthDomainGroups should include(authDomainName)
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        "when the workspace is not shared with them" - {
-          "cannot be seen and is not accessible" in {
-            val user = UserPool.chooseAuthDomainUser
-            implicit val authToken: AuthToken = authTokenDefault
-            withGroup("AuthDomain", List(user.email)) { authDomainName =>
-              withCleanBillingProject(defaultUser) { projectName =>
-                withWorkspace(projectName, "AuthDomainSpec", Set(authDomainName)) { workspaceName =>
-                  withWebDriver { implicit driver =>
-                    withSignIn(user) { workspaceListPage =>
-                      eventually {
-                        workspaceListPage.hasWorkspace(projectName, workspaceName) shouldEqual false
-                      }
-
-                      val workspaceSummaryPage = new WorkspaceSummaryPage(projectName, workspaceName).open
-                      checkWorkspaceFailure(workspaceSummaryPage, projectName, workspaceName)
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        //TCGA controlled access workspaces use-case
-        "when the workspace is shared with the group" - {
-          "can be seen and is accessible" in {
-            val user = UserPool.chooseAuthDomainUser
-            implicit val authToken: AuthToken = authTokenDefault
-            withGroup("AuthDomain", List(user.email)) { groupOneName =>
-              withGroup("AuthDomain", List(user.email)) { groupTwoName =>
-                withCleanBillingProject(defaultUser) { projectName =>
-                  withWorkspace(projectName, "AuthDomainSpec_share", Set(groupOneName, groupTwoName), List(AclEntry(groupNameToEmail(groupOneName), WorkspaceAccessLevel.Reader))) { workspaceName =>
-                    withWebDriver { implicit driver =>
-                      withSignIn(user) { listPage =>
-                        val summaryPage = listPage.enterWorkspace(projectName, workspaceName)
-                        eventually {
-                          summaryPage.readAuthDomainGroups should include(groupOneName)
-                        }
-                        eventually {
-                          summaryPage.readAuthDomainGroups should include(groupTwoName)
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-
-    //MULTI-GROUP AUTH DOMAIN TESTS
-
     "with multiple groups inside of it" - {
+
       "can be created" in {
         val user = UserPool.chooseAuthDomainUser
         implicit val authToken: AuthToken = authTokenDefault
@@ -566,6 +377,7 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
               }
             }
           }
+
           "and given writer access" - {
             "the user has correct permissions" in {
               val user = UserPool.chooseStudent
@@ -588,6 +400,7 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
               }
             }
           }
+
           "when the user is a billing project owner" - {
             "can be seen and is accessible" in {
               val user = UserPool.chooseProjectOwner
@@ -614,6 +427,7 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
             }
           }
         }
+
         "when shared with one of the groups in the auth domain" - {
           "can be seen and is accessible by group member who is a member of both auth domain groups" in {
             val user = UserPool.chooseStudent
@@ -661,6 +475,7 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
             }
           }
         }
+
         "when not shared with them" - {
           "cannot be seen and is not accessible" in {
             val user = UserPool.chooseStudent
@@ -676,6 +491,33 @@ class AuthDomainWorkspaceSpec extends FreeSpec /*with ParallelTestExecution*/ wi
 
                           val workspaceSummaryPage = new WorkspaceSummaryPage(projectName, workspaceName).open
                           checkWorkspaceFailure(workspaceSummaryPage, projectName, workspaceName)
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+
+        //TCGA controlled access workspaces use-case
+        "when the workspace is shared with the group" - {
+          "can be seen and is accessible" in {
+            val user = UserPool.chooseAuthDomainUser
+            implicit val authToken: AuthToken = authTokenDefault
+            withGroup("AuthDomain", List(user.email)) { groupOneName =>
+              withGroup("AuthDomain", List(user.email)) { groupTwoName =>
+                withCleanBillingProject(defaultUser) { projectName =>
+                  withWorkspace(projectName, "AuthDomainSpec_share", Set(groupOneName, groupTwoName), List(AclEntry(groupNameToEmail(groupOneName), WorkspaceAccessLevel.Reader))) { workspaceName =>
+                    withWebDriver { implicit driver =>
+                      withSignIn(user) { listPage =>
+                        val summaryPage = listPage.enterWorkspace(projectName, workspaceName)
+                        eventually {
+                          summaryPage.readAuthDomainGroups should include(groupOneName)
+                        }
+                        eventually {
+                          summaryPage.readAuthDomainGroups should include(groupTwoName)
                         }
                       }
                     }
